13_Multi-Agents-Tool-Calling-Handsoff/requirements.txt
requests
python-dotenv
rich
openai-agents
